Changed p300 bmc from a revision to a board type
The p300 bmc is based on a completely different soc from the rest of the
product line. To better reflect this difference I turned p300 into its
own board. An ideal solution would involve moving the p300 into a
seperate directory, but in the interest of going from 0 -> 1 I made the
smallest diff I could.

Signed-off-by: Daniel Rosen <drosen@tenstorrent.com>