Reset both p300 chips if pcie is already booted
If we detect that pcie has already started its boot when we are
initializing the bmfw we previously only reset the primary chip. This
happens to work on the p1x0 boards because the primary chip is the only
asic, but on the p300 with two asics one asic wouldn't recover post
tt-flash reset.

Signed-off-by: Daniel Rosen <drosen@tenstorrent.com>