tests: tt_boot_fs: choose non-corrupt fw_pack-80.15.0.0
The previous fw_pack-80.14.3.0.fwbundle contained an invalid
spi_addr for the failover image, which was 0x4040. Not only
did this spi address not align properly, but it wasn't even
accurate.

The only reason this was missed in CI was due to a copy-paste
error (fixed in the next commit).

Signed-off-by: Chris Friedt <cfriedt@tenstorrent.com>