boards: tenstorrent: tt_blackhole: reset flash during init sequence
If the bootrom configures the flash to Octal/Quad mode, than the MSPI
NOR driver init will fail. As such, we should reset the flash during
early boot to make sure that it is in a known state. Check the
SPI_DEVICE_ID and SPI_DEVICE_CONFIG regs to determine if this is
required.

Signed-off-by: Daniel DeGrasse <ddegrasse@tenstorrent.com>